Implement `KVM_TDX_INIT_MEM_REGION`
Implements the `KVM_TDX_INIT_MEM_REGION` ioctl for TDX. This will
encrypt a memory continuous region, which correstponds to the
`TDH.MEM.PAGE.ADD` SEAM call. Based on the attributes of the TDVF
section it will also extend the measurement which corresponds to the
`TDH.MR.EXTEND` SEAM call.

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>